doing load contact us

+++ api/routes/pages.js
@@ -2,7 +2,7 @@
 const express = require("express");
 const fs = require("fs");
 
-const { select } = require("sito-node-mysql");
+const { select, update } = require("sito-node-mysql");
 
 // auth
 const { validator } = require("../utils/secure");
@@ -22,10 +22,51 @@ const parseHomeData = async (section, data) => {
       const responsePhoto = await select("config", ["homeContactUsPhoto"]);
       const { rows } = responsePhoto;
       const { homeContactUsPhoto } = rows[0];
-      if (homeContactUsPhoto.length) {
-        // deleting current photo
-        fs.rmSync(`./public${homeContactUsPhoto}`);
+      if (photo)
+        if (homeContactUsPhoto.length) {
+          // deleting current photo
+          fs.rmSync(`./public${homeContactUsPhoto}`);
+        }
+      let thePhoto = "";
+
+      if (photo && photo.photo && photo.photo.blob) {
+        const { blob } = photo.photo;
+        // writing new photo
+        const encoded = Buffer.from(
+          blob.replace(/^data:image\/\w+;base64,/, ""),
+          "base64"
+        );
+        const extension = blob.split(";")[0].split("/")[1];
+        fs.writeFileSync(
+          `./public/images/config/homeContactUsPhoto.${extension}`,
+          encoded
+        );
+        thePhoto = `/images/config/homeContactUsPhoto.${extension}`;
       }
+      // updating photo
+      await update(
+        "config",
+        ["homeContactUsPhoto"],
+        {
+          homeContactUsPhoto: thePhoto,
+        },
+        {}
+      );
+      // updating social media
+      for (const item of socialMedia) {
+        const { id, active } = item;
+        await update(
+          "socialMedia",
+          ["active"],
+          { active },
+          {
+            attribute: "id",
+            value: id,
+            operator: "=",
+          }
+        );
+      }
+
       break;
     }
     default: // hero
@@ -33,7 +74,7 @@ const parseHomeData = async (section, data) => {
   }
 };
 
-pagesRoute.post("save", [validator], async (req, res) => {
+pagesRoute.post("/save", [validator], async (req, res) => {
   const { page, data, user } = req.body;
   try {
     // validating user
